+++ internal/entity/send_request.go
@@ -2,5 +2,5 @@ package entity
 
 type SendCoinRequest struct {
 	ToUser string `json:"toUser" binding:"required"`
-	Amount int    `json:"amount" binding:"required,gt=0"`
+	Amount int64  `json:"amount" binding:"required,gt=0"`
 }
