+++ internal/entity/user.go
@@ -1,8 +1,8 @@
 package entity
 
 type User struct {
-	ID       int    `json:"-" db:"id"`
+	ID       int64  `json:"-" db:"id"`
 	Username string `json:"username" db:"username"`
 	Password string `json:"password" db:"password_hash"`
-	Coins    int    `json:"coins" db:"coins"`
+	Coins    int64  `json:"coins" db:"coins"`
 }
